python-backend/requirements.txt
openai==1.12.0
google-generativeai==0.3.2
anthropic==0.18.1
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0
python-multipart==0.0.9